# Fix performance regression from introduction of Template#disable_tags