… lib/liquid/registers/disabled_tags.rb
… test/integration/registers/disabled_tags_test.rb
… test/unit/registers/disabled_tags_unit_test.rb