@dylanahsmith 2020-08-11
> The liquid-c benchmark seems to indicate the new code is slower but it's all within margin of error.

Oops, it looks like I made a mistake in pasting the benchmark results, so got them mixed up. I've edited the PR description to fix the mistake.  The render line is the one that is relevant and effect is significant is significant at about a 9% improvement. I re-ran the benchmark to verify (got 121.705 i/s on master and 133.317 i/s on the branch)